Add initial-scale=1 to viewport
When initial-scale is undefiend, media query's width always keeps the
value of portrait mode. For example, in case a tablet has 768px x 1024px
resolution, the width is always 768px whether portrait or landscape
mode. Therefore, the contents are enlarged 1.3X (1024 / 768) when in
landscape mode. But initial-scale=1 denies enlarging. See also
<http://webdesignrecipes.com/web-design-for-mobile-with-css3-media-queries/>.

The above web page says that the contents don't fit the window when
iPhone is turned from portrait mode to landscape mode with
initial-scale=1. But I can't confim that with iPhone 5 + iOS 7.1 and
iPad mini Retina + iOS 7.1.